Added files via upload
The product itself does what it intends to, and it works great. The only thing I disliked was how it closed after every selection. 
I removed the lines of code that close the box when finished (to achieve my desired result), but my question is, what would be a better way to keep the box active?